+++ src/progressed/world/blocks/defence/turret/multi/mounts/ChargeMount.java
@@ -6,7 +6,7 @@ import progressed.world.blocks.defence.turret.multi.modules.*;
 public class ChargeMount extends BaseMount{
     public float charge, heat;
 
-    public ChargeMount(ModularTurretBuild parent, BaseModule module, short moduleNumber){
+    public ChargeMount(ModularTurretBuild parent, BaseModule module, int moduleNumber){
         super(parent, module, moduleNumber);
     }
 }
